- Move more bot routes. - Properly revoke admin tokens for previous admins. - Fix error handler.
