+++ src/Routes/bots/_id/decline.ts
@@ -1,0 +1,187 @@
+import { PathRoute } from "../../route.ts";
+import { variables } from "../../../Util/Function/variables.ts";
+import * as permission from "../../../Util/Function/permissions.ts";
+import e from "express";
+import * as userCache from "../../../Util/Services/userCaching.ts";
+import * as botCache from "../../../Util/Services/botCaching.ts";
+import * as Discord from "discord.js";
+import settings from "../../../../settings.json" with { type: "json" };
+import * as discord from "../../../Util/Services/discord.ts";
+import * as functions from "../../../Util/Function/main.ts";
+import { botType } from "../index.ts";
+
+export class GetDeclineBot extends PathRoute<"get"> {
+
+    constructor() {
+        super("get", "/:id/decline", [variables, permission.auth, permission.mod]);
+    }
+
+    async handle(req: e.Request, res: e.Response, next: e.NextFunction) {
+        const bot: delBot = await global.db
+            .collection<delBot>("bots")
+            .findOne({ _id: req.params.id });
+
+        if (!bot)
+            return res.status(404).render("status", {
+                res,
+                title: res.__("common.error"),
+                status: 404,
+                subtitle: res.__("common.error.bot.404"),
+                req,
+                type: "Error"
+            });
+
+        res.locals.premidPageInfo = res.__("premid.bots.decline", bot.name);
+
+        if (bot.status.approved === true)
+            return res.status(400).render("status", {
+                res,
+                title: res.__("common.error"),
+                status: 400,
+                subtitle: res.__("common.error.bot.notInQueue"),
+                req,
+                type: "Error"
+            });
+
+        let redirect = `/bots/${bot._id}`;
+
+        if (req.query.from && req.query.from === "queue")
+            redirect = "/staff/bot_queue";
+
+        res.render("templates/bots/staffActions/remove", {
+            title: res.__("page.bots.decline.title"),
+            icon: "times",
+            subtitle: res.__("page.bots.decline.subtitle", bot.name),
+            req,
+            redirect
+        });
+    }
+
+}
+
+export class PostDeclineBot extends PathRoute<"post"> {
+
+    constructor() {
+        super("post", "/:id/decline", [variables, permission.auth, permission.mod]);
+    }
+
+    async handle(req: e.Request, res: e.Response, next: e.NextFunction) {
+        const bot: delBot | undefined = await global.db
+            .collection<delBot>("bots")
+            .findOne({ _id: req.params.id });
+
+        if (!bot)
+            return res.status(404).render("status", {
+                res,
+                title: res.__("common.error"),
+                status: 404,
+                subtitle: res.__("common.error.bot.404"),
+                req,
+                type: "Error"
+            });
+
+        if (bot.status.approved === true)
+            return res.status(400).render("status", {
+                res,
+                title: res.__("common.error"),
+                status: 400,
+                subtitle: res.__("common.error.bot.notInQueue"),
+                req,
+                type: "Error"
+            });
+
+        if (!req.body.reason && !req.user.db.rank.admin) {
+            return res.status(400).render("status", {
+                res,
+                title: res.__("common.error"),
+                status: 400,
+                subtitle: res.__("common.error.reasonRequired"),
+                req,
+                type: "Error"
+            });
+        }
+
+        await global.db.collection("bots").updateOne(
+            { _id: req.params.id },
+            {
+                $set: {
+                    vanityUrl: "",
+                    lastDenyReason: req.body.reason,
+                    "status.archived": true
+                }
+            }
+        );
+
+        await global.db.collection("users").updateOne(
+            { _id: req.user.id },
+            {
+                $inc: {
+                    "staffTracking.handledBots.allTime.total": 1,
+                    "staffTracking.handledBots.allTime.declined": 1,
+                    "staffTracking.handledBots.thisWeek.total": 1,
+                    "staffTracking.handledBots.thisWeek.declined": 1
+                }
+            }
+        );
+
+        await userCache.updateUser(req.user.id);
+
+        const type = botType(req.body.type);
+
+        await global.db.collection("audit").insertOne({
+            type: "DECLINE_BOT",
+            executor: req.user.id,
+            target: req.params.id,
+            date: Date.now(),
+            reason: req.body.reason || "None specified.",
+            reasonType: type
+        });
+
+        await botCache.updateBot(req.params.id);
+
+        const embed = new Discord.EmbedBuilder();
+        embed.setColor(0x2f3136);
+        embed.setTitle("Reason");
+        embed.setDescription(req.body.reason || "No reason provided.");
+        embed.setURL(`${settings.website.url}/bots/${bot._id}`);
+
+        await discord.channels.logs.send({
+            content: `${settings.emoji.cross} **${functions.escapeFormatting(
+                req.user.db.fullUsername
+            )}** \`(${
+                req.user.id
+            })\` declined bot **${functions.escapeFormatting(bot.name)}** \`(${
+                bot._id
+            })\``,
+            embeds: [embed]
+        });
+
+        const member = await discord.getTestingGuildMember(req.params.id);
+
+        if (member) {
+            await member.kick("Bot's listing has been declined.").catch((e) => {
+                console.error(e);
+            });
+        }
+
+        const owner = await discord.getMember(bot.owner.id);
+        if (owner)
+            owner
+                .send(
+                    `${
+                        settings.emoji.cross
+                    } **|** Your bot **${functions.escapeFormatting(
+                        bot.name
+                    )}** \`(${bot._id})\` has been declined.\n**Reason:** \`${
+                        req.body.reason || "None specified."
+                    }\``
+                )
+                .catch((e) => {
+                    console.error(e);
+                });
+
+        res.redirect("/staff/bot_queue");
+    }
+
+}
+
